update function in my task

+++ WebAPI/Controllers/PremiumPackageController.cs
@@ -1,9 +1,13 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 using Repositories.EntityModels;
 using Repositories.Repository.Interface;
 using Services;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using System.Text;
 using WebAPI.ViewModels;
 
 namespace WebAPI.Controllers
@@ -12,6 +16,7 @@ namespace WebAPI.Controllers
     [ApiController]
     public class PremiumPackageController : ControllerBase
     {
+        private readonly IConfiguration configuration;
         private readonly IPremiumPackageService packageService;
         private readonly IMapper mapper;
 
@@ -19,6 +24,27 @@ namespace WebAPI.Controllers
         {
             this.packageService = packageService;
             this.mapper = mapper;
+        }
+
+        private string GenerateJwtToken(string id, string role)
+        {
+            var jwtSettings = configuration.GetSection("JwtSettings");
+            var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(jwtSettings["SecretKey"]));
+            var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);
+            var claims = new List<Claim>
+            {
+            new Claim(ClaimTypes.NameIdentifier, id),
+            new Claim(ClaimTypes.Role, role)
+        };
+            var token = new JwtSecurityToken(
+                issuer: jwtSettings["Issuer"],
+                audience: jwtSettings["Audience"],
+                claims: claims,
+                expires: DateTime.UtcNow.AddMinutes(30),
+                signingCredentials: creds
+            );
+
+            return new JwtSecurityTokenHandler().WriteToken(token);
         }
 
         [HttpGet]
@@ -68,46 +94,66 @@ namespace WebAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(model.PackageName))
-                    {
-                    return StatusCode(400, new
-                    {
-                        Message = "Package Name cannot empty!!!"
-                    });
-                }
-                else if (model.PackageAmount < 0)
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
                 {
-                    return StatusCode(400, new
+                    if (role == "Admin")
                     {
-                        Message = "PackageAmount cannot small than 0!!!"
-                    });
-                }
-                else if (model.PackageDiscount < 0)
-                {
-                    return StatusCode(400, new
+                        if (string.IsNullOrEmpty(model.PackageName))
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "Package Name cannot empty!!!"
+                            });
+                        }
+                        else if (model.PackageAmount < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageAmount cannot small than 0!!!"
+                            });
+                        }
+                        else if (model.PackageDiscount < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageDiscount cannot small than 0!!!"
+                            });
+                        }
+                        else if (model.PackageMonth < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageMonth cannot small than 0!!!"
+                            });
+                        }
+                        else
+                        {
+                            var pre = mapper.Map<PremiumPackage>(model);
+                            var check = packageService.Add(pre);
+                            return await check ? Ok(new
+                            {
+                                Message = "Add Success"
+                            }) : Ok(new
+                            {
+                                Message = "Add Fail"
+                            });
+                        }
+                    }
+                    else
                     {
-                        Message = "PackageDiscount cannot small than 0!!!"
-                    });
-                }
-                else if (model.PackageMonth < 0)
-                {
-                    return StatusCode(400, new
-                    {
-                        Message = "PackageMonth cannot small than 0!!!"
-                    });
+                        return StatusCode(400, new
+                        {
+                            Status = -1,
+                            Message = "Role Denied"
+                        });
+                    }
                 }
                 else
                 {
-                    var pre = mapper.Map<PremiumPackage>(model);
-                    var check = packageService.Add(pre);
-                    return await check ? Ok(new
-                    {
-                        Message = "Add Success"
-                    }) : Ok(new
-                    {
-                        Message = "Add Fail"
-                    });
+                    return Unauthorized();
                 }
+                
             }
             catch (Exception ex)
             {
@@ -120,46 +166,66 @@ namespace WebAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(model.PackageName))
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
                 {
-                    return StatusCode(400, new
+                    if (role == "Admin")
                     {
-                        Message = "Package Name cannot empty!!!"
-                    });
-                }
-                else if (model.PackageAmount < 0)
-                {
-                    return StatusCode(400, new
+                        if (string.IsNullOrEmpty(model.PackageName))
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "Package Name cannot empty!!!"
+                            });
+                        }
+                        else if (model.PackageAmount < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageAmount cannot small than 0!!!"
+                            });
+                        }
+                        else if (model.PackageDiscount < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageDiscount cannot small than 0!!!"
+                            });
+                        }
+                        else if (model.PackageMonth < 0)
+                        {
+                            return StatusCode(400, new
+                            {
+                                Message = "PackageMonth cannot small than 0!!!"
+                            });
+                        }
+                        else
+                        {
+                            var pre = mapper.Map<PremiumPackage>(model);
+                            var check = packageService.Update(pre);
+                            return await check ? Ok(new
+                            {
+                                Message = "Update Success"
+                            }) : Ok(new
+                            {
+                                Message = "Update Fail"
+                            });
+                        }
+                    }
+                    else
                     {
-                        Message = "PackageAmount cannot small than 0!!!"
-                    });
-                }
-                else if (model.PackageDiscount < 0)
-                {
-                    return StatusCode(400, new
-                    {
-                        Message = "PackageDiscount cannot small than 0!!!"
-                    });
-                }
-                else if (model.PackageMonth < 0)
-                {
-                    return StatusCode(400, new
-                    {
-                        Message = "PackageMonth cannot small than 0!!!"
-                    });
+                        return StatusCode(400, new
+                        {
+                            Status = -1,
+                            Message = "Role Denied"
+                        });
+                    }
                 }
                 else
                 {
-                    var pre = mapper.Map<PremiumPackage>(model);
-                    var check = packageService.Update(pre);
-                    return await check ? Ok(new
-                    {
-                        Message = "Update Success"
-                    }) : Ok(new
-                    {
-                        Message = "Update Fail"
-                    });
+                    return Unauthorized();
                 }
+                
             }
             catch (Exception ex)
             {
@@ -172,23 +238,41 @@ namespace WebAPI.Controllers
         {
             try
             {
-                var pre = packageService.Get(id);
-                if (pre != null)
+                var role = User.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)?.Value;
+                if (!string.IsNullOrEmpty(role))
                 {
-                    var check =  packageService.Delete(id);
-                    return await check ? Ok(new
+                    if (role == "Admin")
                     {
-                        Message = "Delete Success"
-                    }) : Ok(new
+                        var pre = packageService.Get(id);
+                        if (pre != null)
+                        {
+                            var check = packageService.Delete(id);
+                            return await check ? Ok(new
+                            {
+                                Message = "Delete Success"
+                            }) : Ok(new
+                            {
+                                Message = "Delete Fail"
+                            });
+                        }
+                        else
+                        {
+                            return NotFound();
+                        }
+                    }
+                    else
                     {
-                        Message = "Delete Fail"
-                    });
+                        return StatusCode(400, new
+                        {
+                            Status = -1,
+                            Message = "Role Denied"
+                        });
+                    }
                 }
                 else
                 {
-                    return NotFound();
-                }
-               
+                    return Unauthorized();
+                }             
             }
             catch (Exception ex)
             {
